+++ CHANGELOG.md
@@ -6,7 +6,17 @@
 ##### ViewMenu Only
 * Added the `persist` kwarg to `ViewButton`. This prevents link buttons from being disabled/removed when the menu times out or is stopped so they can remain clickable
 
-
+#### Changes
+##### ReactionMenu & ViewMenu
+* The default behavior for the below methods have changed. Previously, using the below methods would return/stop menu's for both `ReactionMenu` & `ViewMenu`. With this update, each method by default now returns or stops menu's according to whichever class the method was invoked from. For example, `ViewMenu.stop_all_sessions()` only stops all `ViewMenu` sessions instead of all `ReactionMenu` sessions and `ViewMenu` sessions. This is done with the addition of the `fixed` parameter for the following methods. If you like the old behavior, simply set the `fixed` parameter to `False`
+  * `.get_all_dm_sessions(fixed=True)`
+  * `.get_all_sessions(fixed=True)`
+  * `.get_session(name: str, fixed=True)`
+  * `.stop_all_sessions(fixed=True)`
+  * `.stop_session(name: str, include_all=False, fixed=True)`
+  * `.get_sessions_count(fixed=True)`
+  
+  With this change, method `.stop_only()` is now deprecated
 
 
 ## v3.0.1 » Feb. 02, 2022